+++ solid_backend/utility/forms.py
@@ -21,19 +21,3 @@ class HasImgForm(forms.ModelForm):
             self.add_error("img_alt", error_message)
 
         super(HasImgForm, self).clean()
-
-
-class DateOrderForm(forms.ModelForm):
-    """
-    Validate whether a validity start date and end date are in the correct order.
-    """
-
-    def clean(self):
-        valid_from = self.cleaned_data.get("valid_from")
-        valid_to = self.cleaned_data.get("valid_to")
-
-        if valid_to is not None and valid_to < valid_from:
-            error_message = "Ensure the end date is equal with or after the start date."
-            self.add_error("valid_to", error_message)
-
-        super(DateOrderForm, self).clean()
